backend/requirements.txt
Django==5.1.5
psycopg2==2.9.10
python-dotenv==1.0.1
numpy==2.2.0
pandas==2.2.3
gunicorn==23.0.0
scikit-learn==1.6.0
django-cors-headers==4.6.0
djangorestframework==3.15.2